Migrations verbetering
flinke verandering in de migrations

+++ app/Http/Controllers/AuthController.php
@@ -35,7 +35,7 @@ class AuthController extends Controller
 
         $user = Auth::user();
 
-        $token = $user->createToken('authToken')->plainTextToken;
+        $token = $user->createToken('authToken', ['*'], now()->addDay())->plainTextToken;
 
         return response()->json([
             'access_token' => $token,
@@ -44,6 +44,11 @@ class AuthController extends Controller
     }
 
     public function logout(Request $request) {
+        if (!$request->user()) {
+            return response()->json([
+                'error' => 'No user logged in',
+            ], 400);
+        }
         $request->user()->tokens()->delete();
 
         return response()->json([
@@ -52,53 +57,5 @@ class AuthController extends Controller
     }
 
     public function register(Request $request) {
-
-        /**
-         * Items to fill with register
-         *
-         * name - max 64 chars
-         * email - just an email
-         * password - plain text password min 8 chars max 64 chars
-         *
-         * optional:
-         * role_id - int from 1 to n (n being the last id of the roles)
-         * department_id - int from 1 to n (n being the last id of the departments)
-        **/
-
-        // Parse the raw JSON data
-        $data = $request->all();
-
-        // Define validation rules
-        $validator = Validator::make($data, [
-            'name' => 'required|string|max:64',
-            'email' => 'required|email|unique:users,email',
-            'password' => 'required|string|min:8|max:64',
-            'role_id' => 'nullable|integer|exists:roles,id',
-            'department_id' => 'nullable|integer|exists:departments,id',
-        ]);
-
-        // Check if the validation fails
-        if ($validator->fails()) {
-            // Return a JSON response with validation errors
-            return response()->json([
-                'errors' => $validator->errors(),
-            ], 422);
-        }
-
-        // Create the user with the validated data
-        $user = User::create([
-            'name' => $data['name'],
-            'email' => $data['email'],
-            'password' => bcrypt($data['password']),
-            'role_id' => $data['role_id'] ?? 1, // Default role ID to 1 if not provided
-            'department_id' => $data['department_id'] ?? null,
-        ]);
-
-
-        // Return a success response
-        return response()->json([
-            'message' => 'User registered successfully!',
-            'user' => $user,
-        ]);
     }
 }
